djangoProject5/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.3.1
idna==3.7
mysql-connector-python==8.4.0
pillow==10.3.0
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
pycryptodome==3.12.0
